backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==25.3.0
banks==2.2.0
bcrypt==4.3.0
beanie==1.29.0
beautifulsoup4==4.13.5
boto3==1.40.18
botocore==1.40.18
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.12
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
griffe==1.14.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langsmith==0.3.42
lazy-model==0.2.0
limits==5.2.0
llama-cloud==0.1.42
llama-cloud-services==0.6.69
llama-index==0.12.52
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.52.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-instrumentation==0.4.1
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.3.0
llama-parse==0.6.69
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
motor==3.7.1
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.2.6
openai==1.104.2
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
platformdirs==4.4.0
prettytable==3.16.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.13.0
PyMuPDF==1.26.0
pypdf==5.9.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2025.9.18
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
striprtf==0.0.26
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
toml==0.10.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
